I was trying to prove something wrong
In the estimate, I was summing over a set. This was not what I intended to prove because the function inside was not injective and was purely wrong.
Instead, we have to sum over lists to allow for duplicate values in the sum. This will make the proofs a bit more tedious.